requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
Django==5.0.6
djangorestframework==3.15.1
exceptiongroup==1.2.1
factory-boy==3.3.0
Faker==25.4.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
iniconfig==2.0.0
msgpack==1.0.8
packaging==24.0
pluggy==1.5.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pytest==8.2.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.4
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.0
urllib3==2.2.1
zope.interface==6.4.post2
